requirements.txt
Flask
openai
gTTS
python-dotenv
requests
google-generativeai